pyproject.toml
[tool.poetry]
name = "search_home"
version = "0.1.0"
description = ""
authors = ["nixissmtp <gleb@simple2b.(none)>"]

[tool.poetry.dependencies]
python = "^3.9"
beautifulsoup4 = "^4.10.0"
pyTelegramBotAPI = "^4.4.0"
python-dotenv = "^0.19.2"
celery = "^5.2.3"
redis = "^4.1.4"
SQLAlchemy = "^1.4.32"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
